feat(kms): port drawing pipeline from pixman to Vulkan (Phase 4.1)
Replaces pixman as the drawing/compositing backend across the KMS
yserver with a Vulkan implementation. All RENDER drawing ops, core
drawing ops, glyph rendering, scanout flipping, mirror management,
and clipping/transform/repeat handling now run through Vulkan
pipelines. The pixman crate is no longer in the workspace dep tree.

## Sub-phases landed

- 4.1.1 — VkContext lifecycle (instance / physical / logical device,
  validation-layer toleration, extension fallbacks, best-effort
  init).
- 4.1.2 — Vulkan-fed scanout: ScanoutBoPool with GBM-allocated bos,
  per-bo VkSemaphore + fence cycle, explicit-fence atomic page-flip
  submission, advancement on pageflip-complete.
- 4.1.3 — Per-window DrawableImage mirrors, damage-driven uploads,
  composite-pass replaces the shadow-blit, AABB recursion-level cull,
  PixmanShadow scanout mode retired.
- 4.1.4 — Drawing-op family port:
  * .1 solid fill via cmd_clear_attachments
  * .2 CopyArea via cmd_copy_image
  * .3 PutImage / GetImage via copy_buffer/image
  * .4 stroke poly ops route through fill pipeline
  * .5 glyph atlas + text-render pipeline
  * .6 RENDER Composite + FillRectangles with affine transform,
    repeat modes, mask, dst-readback for Disjoint/Conjoint, dual-
    source blend for component_alpha; full PictOp enum (Saturate
    + Disjoint 16-27 + Conjoint 32-43)
  * .7 RENDER Trapezoids/Triangles via CPU-rasterised mask + Vk
    Composite; CompositeGlyphs Vk path
  * GC functions (Xor / And / Or / Invert / etc.) via VkLogicOp
    pipeline
- 4.1.5 — Pixman removal. WindowState.image, PixmapState.image,
  CursorState.image, bg_pixmap_image, MirrorUploader, PixmanImage
  newtype — all gone. cpu_types::{Rectangle16, Repeat, PictTransform}
  own the X-protocol data types.

## rendercheck conformance (yserver in vng + lavapipe)

- fill / dcoords / scoords / mcoords / tscoords / tmcoords / blend
  / repeat / bug7366: 100%.
- gradients: 3649/3649 (100%). XRenderColor stops on
  CreateLinearGradient / CreateRadialGradient arrive STRAIGHT (per
  protocol + rendercheck t_gradient.c:119-123); the LUT fill now
  premultiplies.
- triangles: 456/456 (100%). try_vk_render_traps_or_tris resolves
  Drawable + Gradient sources (mirroring the regular Composite
  path including the alpha-vs-no-alpha view selection), wires up
  dst-readback for Disjoint/Conjoint ops, and composites across
  the entire destination for "zero source has effect" ops (Clear,
  Src, In, InReverse, Out, AtopReverse + Disjoint/Conjoint
  variants), matching pixman in
  pixman/pixman-trap.c::get_trap_extents.
- composite / cacomposite: correctness OK; per-op submit +
  queue_wait_idle still pushes them past the 600 s/1800 s budgets.
  Batching is the open follow-up.

## xts protocol-test parity vs the pixman baseline (108e0ad)

| scenario | now PASS | baseline PASS | per-test PASS |
|----------|---------:|--------------:|--------------:|
| Xproto   | 332      | 358           | 91.7% vs 92.0% |
| Xlib3    | 98       | 100           | 61.3% vs 64.1% |

FAIL / UNRES counts match the baseline exactly; the Xproto delta
is entirely UNTESTED tests the per-op cadence ran out of time on
at the same wall-clock budget.

## Misc

- yserver-{e16,fvwm3,wmaker}-xterm-hw recipes use Venus passthrough
  so guests run on the host GPU rather than llvmpipe.
- tools/xts-run.sh locates xts-report via find (the build layout
  shifted; the old hard-coded path swallowed errors silently).
- README updated to reflect the Vulkan backend.

+++ crates/yserver-protocol/src/x11/request_lengths.rs
@@ -677,7 +677,7 @@ pub fn invalid_value(opcode: u8, header_data: u8, body: &[u8]) -> Option<u32> {
         // ConfigureWindow: mask u16 @ body[4..6], values @ body[8].
         // Only stack_mode (bit 6) has a fixed range we enforce here.
         12 if body.len() >= 6 => {
-            let mask = u32::from(read_u16_le(&body[4..6]));
+            let mask = read_u16_le(&body[4..6]);
             validate_value_list(body, 8, mask, cfg_validate_bit)
         }
         // CirculateWindow: header.data = direction ∈ {RaiseLowest=0, LowerHighest=1}.
